requirements/dev.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.9
django-celery-beat==1.6.0
django-celery-results==1.1.2
django-cors-headers==3.2.1
django-extensions==2.2.5
django-timezone-field==3.1
entrypoints==0.3
idna==2.8
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kombu==4.6.7
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
pyrsistent==0.15.6
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.3.11
requests==2.22.0
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
